examples/agents/default/pyproject.toml
[project]
name = "default-agent"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "strands-agents",
    "flask",
    "boto3",
    "gunicorn",
]
